Comment out seutil_relabelto_bin_policy
We comment out the use of the seutil_relabelto_bin_policy call in the
files_relabel_non_auth_files interface. This allows us to set this
interface in a tunable statement, like so:

  seutil_relabelto_bin_policy(foo_t)

  tunable_policy(`foo_relabel_non_auth_files',`
    files_relabel_non_auth_files(foo_t)
  ')

In larger entries, this allows us to have a minimalistic policy (a
domain only allowed to manage and relabel a certain set of file types)
and, through a boolean, enable it to manage and relabel a larger set of
types.