Revert change to fix build
Recently I introduced support for kdevtmpfs managing all device nodes
but this triggers an issue. That method requires the storage module to
be in the base, which pulls in an entire set of other issues, or that
the attributes used by the storage module are moved towards the devices
module (or another module inside base) which also has quite some work on
it.

Going to check with other team first.