Cachefilesd module updates
The module was incomplete because I was not confident about various rules implemented in the Fedora policy.

The files_create_all_files_as_is(cachefilesd_t) interface call was a bad idea. Instead it just needs to
maintain files in the cache with the cachefiles_cache_t type.

The process associated with cachefiles_kernel_t is a kernel thread (kworker)