postmap is a user command
When a postfix admin updates a postfix database, he has to call
"postmap hash:/etc/postfix/databasename" in order to regenerate the
database (in case of a hash database in the example).

To allow postmap to give feedback on errors, grant it access to the user
terminals and private file descriptors of the admin.